refactor(mois): mover endpoint hotmart para controller de mois
