--- backend/ads-service/src/main/java/com/marketinghub/settings/dto/HotmartCollectedProductListResponse.java
@@ -1,9 +1,0 @@
-package com.marketinghub.settings.dto;
-
-import java.util.List;
-
-public record HotmartCollectedProductListResponse(
-        String workspaceId,
-        List<HotmartCollectedProductDto> items
-) {
-}
